enemies now have a coin value that works, fixed a crash caused by calculating effective stats and also made those calculate correctly. Introduced methods for healing, healing in combat, running away from combat, and restoring health upon respawn. Quality of life for user inputs. Turn manager and action bar can now handle player death. Player character now earns coins for killing an enemy. Player can now spend coins to upgrade stats upon death. Weapon dexterity scalar now applies properly

+++ src/NPC.java
@@ -3,7 +3,7 @@ public class NPC {
     double attack = 0.0;
     double defense = 0.0;
     double dexterity = 0.0;
-    double value = 0.0;
+    int value = 0;
     Item tradeItem;
     WorldGenerator.WorldType worldType;
     public enum Rarity {
@@ -209,7 +209,7 @@ public class NPC {
                     enemyType = EnemyType.DRAGON;
             }
             setRarity(character.getEffectiveLuck());
-            value = rarity.baseValue / ((NumberProcessor.getRandom(rarity.valueScalar,100))/100.0);
+            value = (int) (rarity.baseValue / ((NumberProcessor.getRandom(rarity.valueScalar,100))/100.0));
             attack = (NumberProcessor.getRandom(rarity.statsScalarLow,rarity.statsScalarHigh)) * enemyType.attackScalar;
             defense = (NumberProcessor.getRandom(rarity.statsScalarLow,rarity.statsScalarHigh)) * enemyType.defenseScalar;
             dexterity = (NumberProcessor.getRandom(20,100)) * enemyType.dexScalar;
